face_recognition/camera/requirements.txt
numpy==1.21.6
pandas==1.2.4
imutils==0.5.4
opencv-python==4.0.1.24
requests==2.27.1
Pillow==9.4.0
loguru==0.6.0
redis==3.5.3
Jetson.GPIO==2.1.0
